Move to header only when src does nothing
